Extract provider connect state hook
